+++ src/msystem.jl
@@ -1,28 +1,22 @@
 """
 Info about molecular system (geometry/basis).
+
+`Atom` and `FlexibleSystem` from `AtomsBase` package are used
+for the atoms and the molecular system, respectively.
 """
 module MSystem
+using Unitful, UnitfulAtomic
+using AtomsBase
 using Printf
 using DocStringExtensions
 using ..ElemCo.ECInts
 using ..ElemCo.Constants
-export MSys, ms_exists, Basis, ACenter, genxyz, nuclear_repulsion, bond_length, electron_distribution
+export parse_geometry, system_exists, genxyz, nuclear_repulsion, bond_length, electron_distribution
 export guess_nelec, guess_norb, guess_ncore
 export generate_basis
 
 include("elements.jl")
 include("minbas.jl")
-
-"""
-Basis set
-
-$(FIELDS)
-"""
-struct Basis
-  """ e.g., vdz, cc-pvdz, aug-cc-pvdz, cc-pvdz-jkfit. """
-  name::String
-  # coeffs, exponents..
-end
 
 """ 
     genbasis4element(basis::Dict,elem::AbstractString)
@@ -30,7 +24,7 @@ end
   Set element specific basis from, e.g., Dict("ao"=>"cc-pVDZ; o=aug-cc-pVDZ","jkfit"=>"cc-pvdz-jkfit")
 """
 function genbasis4element(basis::Dict,elem::AbstractString)
-  elembasis = Dict{String,Basis}()
+  elembasis = Dict{String,String}()
   elemUP = uppercase(elem)
   for (type,name) in basis
     names = strip.(split(name,[';',',']))
@@ -53,29 +47,11 @@ function genbasis4element(basis::Dict,elem::AbstractString)
         end
       end
     end
-    elembasis[type] = Basis(elbas)
+    elembasis[type] = elbas
   end
   return elembasis
 end
 
-"""
-Atomic center
-
-$(FIELDS)
-"""
-mutable struct ACenter
-  """name (as defined in input)."""
-  name::String
-  """nuclear charge (can be changed...)."""
-  nuccharge::Float64
-  """coordinates in bohr."""
-  coord::AbstractArray{Float64,1}
-  """various basis sets (`"ao"`, `"mp2fit"`, `"jkfit"`)."""
-  basis::Dict{String,Basis}
-end
-
-Base.show(io::IO, val::ACenter) = print(io, val.name, " ", val.coord[1], " ", val.coord[2], " ", val.coord[3]) 
-  
 """ 
     element_name(name::AbstractString)
 
@@ -83,6 +59,15 @@ Base.show(io::IO, val::ACenter) = print(io, val.name, " ", val.coord[1], " ", va
 """
 function element_name(name::AbstractString)
   return titlecase(rstrip(name,['0','1','2','3','4','5','6','7','8','9']),strict=true)
+end
+
+""" 
+    element_name(atom::Atom)
+
+  Return element name without numbers.
+"""
+function element_name(atom::Atom)
+  return element_name(string(atomic_symbol(atom)))
 end
 
 """ 
@@ -95,109 +80,109 @@ function element_NAME(name::AbstractString)
 end
 
 """ 
-    a2b(vals,skip)
+    element_NAME(atom::Atom)
 
-  Transform from angstrom to bohr (if `skip`: no transformation). 
+  Return element name in all caps and without numbers.
 """
-function a2b(vals,skip)
-  if skip
-    return vals
-  else
-    return vals/Constants.BOHR2ANGSTROM
-  end
-end
-
-""" 
-    b2a(vals,skip)
-
-  Transform from bohr to angstrom (if `skip`: no transformation). 
-"""
-function b2a(vals,skip)
-  if skip
-    return vals
-  else
-    return vals*Constants.BOHR2ANGSTROM
-  end
+function element_NAME(atom::Atom)
+  return element_NAME(string(atomic_symbol(atom)))
 end
 
 """
-    try2create_acenter(line::AbstractString, basis::Dict, bohr = true)
+    try2create_atom(line::AbstractString, basis::Dict, unit=u"bohr")
 
-  Create ACenter from a line `<Atom> x y z`. 
+  Create `Atom` from a line `<Atom> x y z`. 
   
-  If !bohr: coords are in angstrom.
+  `unit` is the unit of the coordinates.
   Returns the center and a bool success variable. 
   If the line has a different format: return dummy center and false.
 """
-function try2create_acenter(line::AbstractString, basis::Dict, bohr = true)
+function try2create_atom(line::AbstractString, basis::Dict, unit=u"bohr")
   coords = split(line)
   if length(coords) == 4
     xcoord = tryparse(Float64,coords[2])
     ycoord = tryparse(Float64,coords[3])
     zcoord = tryparse(Float64,coords[4])
     if !isnothing(xcoord) && !isnothing(ycoord) && !isnothing(zcoord)
-      basis = genbasis4element(basis,coords[1])
-      charge = nuclear_charge_of_center(element_NAME(coords[1]))
-      return ACenter(coords[1],charge,a2b([xcoord,ycoord,zcoord],bohr),basis), true
+      basis4a = genbasis4element(basis,coords[1])
+      anum = nuclear_charge_of_center(element_NAME(coords[1]))
+      return Atom(anum, [xcoord,ycoord,zcoord]*unit, atomic_symbol=Symbol(coords[1]), basis=basis4a), true
     end
   end
   # not a center
-  return ACenter("",0.0,[],Dict{String,Basis}()), false
+  return Atom(1, atomic_symbol=:X, [0.0,0.0,0.0]u"bohr"), false
 end
 
 """
-Geometry and basis set for each element name in the geometry
+    is_dummy(ac::Atom)
 
-$(FIELDS)
+  Check whether the atom is a dummy atom.
 """
-mutable struct MSys
-  """array of atomic centers."""
-  atoms::AbstractArray{ACenter,1}
-  function MSys()
-    new(ACenter[])
-  end
-  function MSys(xyz::AbstractString, basis::Dict)
-    xyz_lines = strip.(split(xyz,"\n"))
-    if length(xyz_lines) == 0
-      error("Empty geometry im MSys")
-    end
-    array_of_atoms, badline = parse_xyz_geometry(xyz_lines, basis)
-    if !isempty(badline) && length(xyz_lines) == 1 && isfile(xyz_lines[1])
-      filename = xyz_lines[1]
-      xyz_lines = strip.(readlines(filename))
-      if length(xyz_lines) == 0
-        error("Empty geometry file $filename im MSys")
-      end
-      array_of_atoms, badline = parse_xyz_geometry(xyz_lines, basis)
-    end
-    if !isempty(badline)
-      error("Unsupported xyz line $badline")
-    end
-    new(array_of_atoms)
-  end
+function is_dummy(ac::Atom)
+  return haskey(ac, :dummy)
 end
 
-function Base.show(io::IO, val::MSys) 
-  for atom in val.atoms
-    println(io, atom)
+"""
+    set_dummy!(ac::Atom)
+
+  Set the atom as a dummy atom.
+"""
+function set_dummy!(ac::Atom)
+  ac.data[:dummy] = true
+end
+
+"""
+    unset_dummy!(ac::Atom)
+
+  Unset the atom as a dummy atom.
+"""
+function unset_dummy!(ac::Atom)
+  delete!(ac.data, :dummy)
+end
+
+"""
+    parse_geometry(geometry::AbstractString, basis::Dict)
+
+  Parse geometry `geometry` and return `FlexibleSystem` object.
+  The geometry can be in xyz format or in a file.
+  The basis set can be defined for each element in the geometry.
+"""
+function parse_geometry(geometry::AbstractString, basis::Dict)
+  geom_lines = strip.(split(geometry,"\n"))
+  if length(geom_lines) == 0
+    error("Empty geometry in parse_geometry!")
   end
+  array_of_atoms, badline = parse_xyz_geometry(geom_lines, basis)
+  if !isempty(badline) && length(geom_lines) == 1 && isfile(geom_lines[1])
+    filename = geom_lines[1]
+    geom_lines = strip.(readlines(filename))
+    if length(geom_lines) == 0
+      error("Empty geometry file $filename in parse_geometry!")
+    end
+    array_of_atoms, badline = parse_xyz_geometry(geom_lines, basis)
+  end
+  if !isempty(badline)
+    error("Unsupported geometry line $badline")
+  end
+  return isolated_system(array_of_atoms)
 end
 
 """
     parse_xyz_geometry(xyz_lines::AbstractArray, basis::Dict)
 
   Parse xyz geometry `xyz_lines` stored as a vector of strings.
-  Return array of atomic centers and an empty string in case of success.
+  Return array of `Atom`s and an empty string in case of success.
 
   Empty lines are skipped.
-  If the line is `bohr` or `angstrom`: change the units.
-  If the first line is a number: assume xyz format and skip the second line.
+  The default units are bohr. If the line is `bohr` or `angstrom`: change the units.
+  If the first line is a number: assume xyz format and skip the second line
+  (in this case, the default units are angstroms).
   If parsing fails: return empty array and the line that failed.
 """
 function parse_xyz_geometry(xyz_lines::AbstractArray, basis::Dict)
-  array_of_atoms = ACenter[]
+  array_of_atoms = Atom[]
   badline = ""
-  bohr = true
+  unit = u"bohr"
   firstline = true
   xyz_format = false
   for line in xyz_lines
@@ -206,14 +191,14 @@ function parse_xyz_geometry(xyz_lines::AbstractArray, basis::Dict)
       firstline = false
       continue
     end
-    ac, success = try2create_acenter(line,basis,bohr)
+    ac, success = try2create_atom(line, basis, unit)
     if success
       push!(array_of_atoms,ac)
     elseif line == "bohr"
-      bohr = true
+      unit = u"bohr"
       success = true
     elseif line == "angstrom"
-      bohr = false
+      unit = u"angstrom"
       success = true
     elseif line == ""
       # skip
@@ -223,7 +208,7 @@ function parse_xyz_geometry(xyz_lines::AbstractArray, basis::Dict)
       natoms = tryparse(Int64,line)
       xyz_format = !isnothing(natoms)
       if xyz_format
-        bohr = false # assume angstrom in xyz format
+        unit = u"angstrom" # assume angstrom in xyz format
         # skip the second line
         continue
       end
@@ -238,57 +223,62 @@ function parse_xyz_geometry(xyz_lines::AbstractArray, basis::Dict)
 end
 
 """ 
-    ms_exists(ms::MSys)
+    system_exists(ms::AbstractSystem)
 
   Check whether the system is not empty.
 """
-function ms_exists(ms::MSys)
-  return length(ms.atoms) > 0
+function system_exists(ms::AbstractSystem)
+  return length(ms) > 0
 end
 
 """ 
-    genxyz(ac::ACenter; bohr=true)
+    genxyz(ac::Atom; unit=u"angstrom")
 
   Generate xyz string with element without numbers.
-  If bohr: return in coordinates in bohr.
 """
-function genxyz(ac::ACenter; bohr=true)
-  name = element_name(ac.name)
-  return @sprintf("%-3s %16.10f %16.10f %16.10f",name,b2a(ac.coord[1],bohr),b2a(ac.coord[2],bohr),b2a(ac.coord[3],bohr)) 
+function genxyz(ac::Atom; unit=u"angstrom")
+  name = element_name(ac)
+  coord = uconvert.(unit, ac.position)/unit
+  return @sprintf("%-3s %16.10f %16.10f %16.10f",name,coord[1],coord[2],coord[3]) 
 end
 
 """ 
-    genxyz(ms::MSys; bohr=true)
+    genxyz(ms::AbstractSystem; unit=u"angstrom")
 
   Generate xyz string with elements without numbers.
-  If bohr: return in coordinates in bohr
 """
-function genxyz(ms::MSys; bohr=true)
-  return join([genxyz(at,bohr=bohr) for at in ms.atoms],"\n")
+function genxyz(ms::AbstractSystem; unit=u"angstrom")
+  return join([genxyz(at; unit) for at in ms],"\n")
 end
 
 """
-    bond_length(cen1::ACenter, cen2::ACenter)
+    bond_length(cen1::Atom, cen2::Atom)
 
-  Calculate bond length between two centers.
+  Calculate bond length in bohr between two centers.
 """
-function bond_length(cen1::ACenter, cen2::ACenter)
-  return sqrt(sum(abs2,(cen1.coord-cen2.coord)))
+function bond_length(cen1::Atom, cen2::Atom)
+  return sqrt(sum(abs2,uconvert.(u"bohr", cen1.position-cen2.position)/u"bohr"))
 end
 
 """ 
-    nuclear_repulsion(ms::MSys)
+    nuclear_repulsion(ms::AbstractSystem)
 
   Calculate nuclear repulsion energy.
 """
-function nuclear_repulsion(ms::MSys)
+function nuclear_repulsion(ms::AbstractSystem)
   enuc = 0.0
-  for i=2:length(ms.atoms)
-    at1 = ms.atoms[i]
-    z1 = at1.nuccharge
-    for j=1:i-1
-      at2 = ms.atoms[j]
-      z2 = at2.nuccharge
+  for i = 2:length(ms)
+    at1 = ms[i]
+    if is_dummy(at1)
+      continue
+    end
+    z1 = atomic_number(at1)
+    for j = 1:i-1
+      at2 = ms[j]
+      if is_dummy(at2)
+        continue
+      end
+      z2 = atomic_number(at2)
       enuc += z1*z2/bond_length(at1,at2)
     end
   end
@@ -296,24 +286,24 @@ function nuclear_repulsion(ms::MSys)
 end
 
 """ 
-    guess_nelec(ms::MSys)
+    guess_nelec(ms::AbstractSystem)
 
   Guess the number of electrons in the neutral system.
 """
-function guess_nelec(ms::MSys)
+function guess_nelec(ms::AbstractSystem)
   nelec = 0
-  for at in ms.atoms
-    nelec += nuclear_charge_of_center(element_NAME(at.name))
-  end
+  for at in ms if !is_dummy(at)
+    nelec += atomic_number(at)
+  end end
   return nelec
 end
 
 """
-    guess_nalpha(ms::MSys)
+    guess_nalpha(ms::AbstractSystem)
 
   Guess the number of alpha electrons in the neutral system.
 """
-function guess_nalpha(ms::MSys)
+function guess_nalpha(ms::AbstractSystem)
   nelec = guess_nelec(ms)
   if nelec % 2 == 0
     return nelec ÷ 2
@@ -323,11 +313,11 @@ function guess_nalpha(ms::MSys)
 end
 
 """
-    guess_nbeta(ms::MSys)
+    guess_nbeta(ms::AbstractSystem)
 
   Guess the number of beta electrons in the neutral system.
 """
-function guess_nbeta(ms::MSys)
+function guess_nbeta(ms::AbstractSystem)
   nelec = guess_nelec(ms)
   if nelec % 2 == 0
     return nelec ÷ 2
@@ -337,37 +327,57 @@ function guess_nbeta(ms::MSys)
 end
 
 """
-    guess_nocc(ms::MSys)
+    guess_nocc(ms::AbstractSystem)
 
   Guess the number of alpha and beta occupied orbitals in the neutral system.
 """
-function guess_nocc(ms::MSys)
+function guess_nocc(ms::AbstractSystem)
   nalpha = guess_nalpha(ms)
   nbeta = guess_nbeta(ms)
   return nalpha, nbeta
 end
 
 """
-    guess_norb(ms::MSys)
+    basis_name_AB(atoms, type="ao")
+
+  Return the name of the basis set.
+  `atoms` can be a single atom `::Atom` or a system `::AbstractSystem`.
+"""
+function basis_name_AB(atoms, type="ao")
+  if haskey(atoms, :basis)
+    return lowercase(atoms[:basis][type])
+  else
+    return "unknown"
+  end
+end
+
+"""
+    guess_norb(ms::AbstractSystem)
 
   Guess the number of orbitals in the system.
 """
-function guess_norb(ms::MSys)
+function guess_norb(ms::AbstractSystem)
     # TODO: use element-specific basis!
-  aobasis = lowercase(ms.atoms[1].basis["ao"].name)
-  bao = BasisSet(aobasis,genxyz(ms,bohr=false))
+  aobasis = basis_name_AB(ms, "ao")
+  if aobasis == "unknown"
+    aobasis = basis_name_AB(ms[1], "ao")
+  end
+  if aobasis == "unknown"
+    error("Basis set not defined!")
+  end
+  bao = BasisSet(aobasis, genxyz(ms))
   return bao.nbas 
 end
 
 """
-    guess_ncore(ms::MSys, coretype::Symbol=:large)
+    guess_ncore(ms::AbstractSystem, coretype::Symbol=:large)
 
   Guess the number of core orbitals in the system.
 
   `coretype` as in [`ncoreorbs`](@ref).
 """
-function guess_ncore(ms::MSys, coretype::Symbol=:large)
-  return sum([ncoreorbs(element_NAME(at.name),coretype) for at in ms.atoms])
+function guess_ncore(ms::AbstractSystem, coretype::Symbol=:large)
+  return sum([ncoreorbs(element_NAME(at),coretype) for at in ms if !is_dummy(at)])
 end
 
 """ 
@@ -377,29 +387,32 @@ end
   
   Number of orbitals in the minimal basis set has to be specified in `minbas.jl`.
 """
-function electron_distribution(ms::MSys, minbas::AbstractString)
+function electron_distribution(ms::AbstractSystem, minbas::AbstractString)
   eldist = Float64[]
-  for at in ms.atoms
-    elnam = element_NAME(at.name)
-    nnum = nuclear_charge_of_center(elnam)
-    if nnum == 0
-      continue
-    end
+  for at in ms if !is_dummy(at)
+    elnam = element_NAME(at)
+    nnum = atomic_number(at)
     eldist = vcat(eldist,electron_distribution(elnam,nshell4l_minbas(nnum,uppercase(minbas))))
-  end
+  end end
   return eldist
 end
 
 """
-    generate_basis(ms::MSys, type = "ao")
+    generate_basis(ms::AbstractSystem, type = "ao")
 
   Generate basis sets for integral calculations.
   `type` can be `"ao"`, `"mp2fit"` or `"jkfit"`.
 """
-function generate_basis(ms::MSys, type = "ao")
+function generate_basis(ms::AbstractSystem, type = "ao")
   # TODO: use element-specific basis!
-  basis_name = lowercase(ms.atoms[1].basis[type].name)
-  basis = BasisSet(basis_name,genxyz(ms,bohr=false))
+  basis_name = basis_name_AB(ms, type)
+  if basis_name == "unknown"
+    basis_name = basis_name_AB(ms[1], type)
+  end
+  if basis_name == "unknown"
+    error("Basis set not defined!")
+  end
+  basis = BasisSet(basis_name,genxyz(ms))
   return basis
 end